requirements.txt
jinja2~=2.11.3
Pillow~=8.1.0
